Refactoring and linting error fix for institution
Refactoring of application, content creator and institution and linting fix for institution
