+++ web/src/features/user/types/Institution.ts
@@ -3,5 +3,5 @@ export interface Institution {
   institutionName: string;
   domain: string;
   secondaryDomain: string;
-  [key: string]: any;
+  [key: string]: unknown;
 }
