+++ web/src/unplaced/Admin/Institutions/Actions/AddInstitutionsButton.tsx
@@ -3,7 +3,7 @@ import { useForm, SubmitHandler } from "react-hook-form";
 import { useNavigate } from "react-router-dom";
 import { toast } from "react-toastify";
 
-import { Institution } from "@/features/user/types/Institution";
+import { Institution } from "@/user/types/institution.ts";
 import GenericModalComponent from "@/shared/components/GenericModalComponent";
 import { useNotifications } from "@/shared/context/NotificationContext";
 import { useApi } from "@/shared/hooks/useAPI";
