+++ web/src/unplaced/Admin/Institutions/Actions/DeleteInstitutionsButton.tsx
@@ -4,7 +4,7 @@ import { mdiDelete } from "@mdi/js";
 import { toast } from "react-toastify";
 
 import { getUserToken } from "@/features/auth/lib/userInfo";
-import { Institution } from "@/features/user/types/Institution";
+import { Institution } from "@/user/types/institution.ts";
 import GenericModalComponent from "@/shared/components/GenericModalComponent";
 import { useNotifications } from "@/shared/context/NotificationContext";
 import { useApi } from "@/shared/hooks/useAPI";
