+++ web/src/unplaced/Admin/Users/Actions/ViewUserButton.tsx
@@ -8,8 +8,8 @@ import { getUserToken } from "@/auth/lib/userInfo";
 import { useApi } from "@/shared/hooks/useAPI";
 import AdminServices from "@/unplaced/services/admin.services";
 import AuthServices from "@/unplaced/services/auth.services";
-import { Application } from "@/user/types/Application";
-import { ContentCreator } from "@/user/types/ContentCreator";
+import { Application } from "@/user/types/application.ts";
+import { ContentCreator } from "@/user/types/content-creator.ts";
 import { User } from "@/user/types/User";
 
 import UserDetailsModal from "../DetailsModalUser";
