+++ web/src/unplaced/Admin/Users/UsersTableAdmin.tsx
@@ -12,7 +12,7 @@ import { getUserToken } from "@/auth/lib/userInfo";
 import Loading from "@/shared/components/Loading";
 import AdminServices from "@/unplaced/services/admin.services";
 import ViewUserButton from "@/unplaced/ViewUserButton";
-import { ContentCreator } from "@/user/types/ContentCreator";
+import { ContentCreator } from "@/user/types/content-creator.ts";
 import { User } from "@/user/types/User";
 
 import AdminToggleButton from "../AdminToggle";
