+++ web/src/unplaced/Application.tsx
@@ -10,7 +10,7 @@ import Motivation from "@/user/components/motivation.tsx";
 
 import AcademicExperienceForm from "../features/user/components/academic-experience-form";
 import ProfessionalExperienceForm from "../features/user/components/ProfessionalExperience";
-import { NewApplication } from "../features/user/types/Application";
+import { NewApplication } from "@/user/types/application.ts";
 import GenericModalComponent from "../shared/components/GenericModalComponent";
 import { useApi } from "../shared/hooks/useAPI";
