+++ web/src/unplaced/services/auth.services.ts
@@ -1,8 +1,8 @@
 import axios from "axios";
 
 import { BACKEND_URL } from "@/shared/config/environment";
-import { Application } from "@/user/types/Application";
-import { Institution } from "@/user/types/Institution";
+import { Application } from "@/user/types/application.ts";
+import { Institution } from "@/user/types/institution.ts";
 import { User } from "@/user/types/User";
 import { getUserToken } from "@/auth/lib/userInfo";
